requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
cerebras-cloud-sdk 